Fixed another issue with sources.
I've fixed the "Each child in a list should have a unique 'key' prop" error in the SourcesList component. The issue was that the code was using source.id as the key for each list item, but the source objects don't have an id property.

Here's what I did:

Changed the key in the list items from key={source.id} to key={${source.url}-${index}}
Added the index parameter to the map function to create a unique key for each item
This change ensures that each item in the list has a unique key, which is required by React for efficient rendering and updates. Using a combination of the URL and index ensures uniqueness even if there are duplicate URLs in the sources list.

The error should now be resolved, and you shouldn't see the "Each child in a list should have a unique 'key' prop" warning in the console anymore.